pins refactor
