+++ src/client/components/Story/Story.js
@@ -1,4 +1,4 @@
-import { map, isEmpty, get, toLower, has, some } from 'lodash';
+import { map, isEmpty, get, toLower, has } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { ReactSVG } from 'react-svg';
@@ -27,17 +27,10 @@ import DMCARemovedMessage from './DMCARemovedMessage';
 import ObjectAvatar from '../ObjectAvatar';
 import PostedFrom from './PostedFrom';
 import WeightTag from '../WeightTag';
-import {
-  getAppendData,
-  getObjectName,
-  isObjectReviewTab,
-} from '../../../common/helpers/wObjectHelper';
+import { getObjectName, isObjectReviewTab } from '../../../common/helpers/wObjectHelper';
 import { guestUserRegex } from '../../../common/helpers/regexHelpers';
 
 import './Story.less';
-import { getAppendDownvotes, getAppendUpvotes } from '../../../common/helpers/voteHelpers';
-import { getMinRejectVote, getUpdateByBody } from '../../../waivioApi/ApiClient';
-import { objectFields } from '../../../common/constants/listOfFields';
 
 @injectIntl
 @withRouter
@@ -50,6 +43,7 @@ class Story extends React.Component {
     postState: PropTypes.shape().isRequired,
     rewardFund: PropTypes.shape().isRequired,
     defaultVotePercent: PropTypes.number.isRequired,
+    userVotingPower: PropTypes.number,
     showNSFWPosts: PropTypes.bool.isRequired,
     pendingLike: PropTypes.bool,
     isThread: PropTypes.bool,
@@ -62,7 +56,6 @@ class Story extends React.Component {
     isAuthUser: PropTypes.bool,
     history: PropTypes.shape(),
     wobject: PropTypes.shape(),
-    locale: PropTypes.string,
     showPostModal: PropTypes.func,
     votePost: PropTypes.func,
     toggleBookmark: PropTypes.func,
@@ -70,10 +63,7 @@ class Story extends React.Component {
     editPost: PropTypes.func,
     followUser: PropTypes.func,
     unfollowUser: PropTypes.func,
-    appendObject: PropTypes.func,
-    userVotingPower: PropTypes.number,
-    voteAppends: PropTypes.func,
-    setPinnedPostsUrls: PropTypes.func,
+    handlePinPost: PropTypes.func,
     pinnedPostsUrls: PropTypes.arrayOf(PropTypes.string),
     push: PropTypes.func,
     pendingFlag: PropTypes.bool,
@@ -306,77 +296,6 @@ class Story extends React.Component {
       hiddenStoryPreviewMessage
     );
   }
-  handlePinPost = async () => {
-    const {
-      post,
-      voteAppends,
-      pinnedPostsUrls,
-      setPinnedPostsUrls,
-      user,
-      match,
-      wobject,
-      userVotingPower,
-      locale,
-      appendObject,
-    } = this.props;
-    const currUpdate = await getUpdateByBody(
-      wobject.author_permlink,
-      'pin',
-      'en-US',
-      `${post.author}/${post.permlink}`,
-    );
-    const upVotes = currUpdate?.active_votes && getAppendUpvotes(currUpdate?.active_votes);
-    const isLiked = currUpdate?.isLiked || some(upVotes, { voter: user.name });
-    const downVotes = getAppendDownvotes(currUpdate?.active_votes);
-    const isReject = currUpdate?.isReject || some(downVotes, { voter: user.name });
-    let voteWeight;
-
-    if (pinnedPostsUrls.includes(post.url)) {
-      setPinnedPostsUrls(pinnedPostsUrls.filter(p => p !== post.url));
-      if (post?.currentUserPin) {
-        if (isReject) voteWeight = 0;
-        else {
-          voteWeight =
-            isEmpty(upVotes) || (isLiked && upVotes?.length === 1)
-              ? 1
-              : (
-                  await getMinRejectVote(
-                    user.name,
-                    currUpdate?.author,
-                    currUpdate?.permlink,
-                    match.params.name,
-                  )
-                )?.result;
-        }
-        voteAppends(currUpdate?.author, currUpdate?.permlink, voteWeight, 'pin', false, 'pin');
-      }
-    } else if (currUpdate.message) {
-      setPinnedPostsUrls([...pinnedPostsUrls, post.url]);
-      const pageContentField = {
-        name: objectFields.pin,
-        body: `${post.author}/${post.permlink}`,
-        locale,
-      };
-
-      const bodyMessage = `@${user.name} pinned post: author: ${post.author}, permlink: ${post.permlink}`;
-      const postData = getAppendData(user.name, wobject, bodyMessage, pageContentField);
-
-      appendObject(postData, { votePercent: userVotingPower, isLike: true, isObjectPage: true });
-    } else {
-      setPinnedPostsUrls([...pinnedPostsUrls, post.url]);
-      voteAppends(
-        currUpdate.author,
-        currUpdate.permlink,
-        userVotingPower,
-        'pin',
-        false,
-        'pin',
-        null,
-        false,
-        true,
-      );
-    }
-  };
 
   render() {
     const {
@@ -400,6 +319,8 @@ class Story extends React.Component {
       match,
       wobject,
       isAuthUser,
+      handlePinPost,
+      userVotingPower,
     } = this.props;
     const isObjectPage = isObjectReviewTab(wobject, match) && isAuthUser;
     const currentUserPin = pinnedPostsUrls.includes(post.url);
@@ -504,7 +425,16 @@ class Story extends React.Component {
                           className={currentUserPin ? 'pin-website-color' : pinClassName}
                           wrapper="span"
                           src="/images/icons/pin-outlined.svg"
-                          onClick={this.handlePinPost}
+                          onClick={() =>
+                            handlePinPost(
+                              post,
+                              pinnedPostsUrls,
+                              user,
+                              match,
+                              wobject,
+                              userVotingPower,
+                            )
+                          }
                         />
                       </Tooltip>
                     )}
